refactor: split core model internals
